Fix timestep dtype in load_inputs for HunyuanVideo15
HunyuanVideo15TimeEmbedding.forward casts the sinusoidal projection
with .to(dtype=timestep.dtype), so passing torch.long causes the float
embedding to be converted to Long before it reaches the linear layer,
raising a dtype mismatch error.

Pass the timestep as float32 to avoid this diffusers 0.37.1 behaviour.

+++ hunyuan_video_1_5_i2v_720p_gguf/pytorch/loader.py
@@ -163,7 +163,10 @@ class ModelLoader(ForgeModel):
             config.image_embed_dim,
             dtype=dtype,
         )
-        timestep = torch.tensor([500], dtype=torch.long).expand(batch_size)
+        # Use float32 timestep: HunyuanVideo15TimeEmbedding.forward casts the
+        # sinusoidal projection to timestep.dtype, so Long would produce a Long
+        # tensor that cannot be multiplied by bfloat16 weights.
+        timestep = torch.tensor([500.0], dtype=torch.float32).expand(batch_size)
 
         return {
             "hidden_states": hidden_states,
